backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
psycopg2-binary
django-cors-headers==4.3.1
anthropic==0.20.0
python-dotenv==1.0.0
gunicorn==21.2.0